Improve error handling in HttpClient
Make sure that after an error callback have been made on a
HttpClientConnection no more callbacks will happen. To ensure that the
HTTP parser is canceled to avoid that it continues to process whatever
data was in its internal buffer.

Only ever call one error callback. If the client response input stream
as a n error callback use that otherwise use the one on the
HttpClientConnection object.

R=ager@google.com

BUG=dart:7013, dart:7014

Review URL: https://codereview.chromium.org//11434034

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@15570 260f80e4-7a28-3924-810f-c04153c831b5