Fix extraction of environment variables on Windows.
Environment variables were extracted using GetEnvironmentStringsA
and that does not play nice with for instance Danish letters
in environment variables.

This causes problems for pub if the current user has an
internationalized name.

This is the first patch in a series that will switch to the
unicode APIs on Windows.

R=sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org//11529008

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@15977 260f80e4-7a28-3924-810f-c04153c831b5