Fixing callbacks test to pass in checked mode.
I've been running all the html tests in checked mode locally, and for some reason this one just started failing because it was getting a fractional time.

The signature of the callback takes a num for the time, so the test should use a num rather than an int.

BUG=

Review URL: https://codereview.chromium.org//11639006

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@16294 260f80e4-7a28-3924-810f-c04153c831b5