Fixing the DOM event streams test to work on FF and IE.
I was using the focus event since it's supposed to be synchronous and it's nice to test with a browser-generated event for completeness, but it turns out the IE and FF<18 implement this as an async event (FF 18 appears to be correct).

Just changed the test to manually trigger the event which is a fairly close approximation.

BUG=7888

Review URL: https://codereview.chromium.org//11875022

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@17102 260f80e4-7a28-3924-810f-c04153c831b5