dart2js: Avoid one-character names in minifier for stability
This reduces the size of vcdiff output from version to version of your program.
Also sort some names in the interceptor lists, so they have the same order from
run to run.  The net result is a 1% increase in minified size, and a 66%
decrease in vcdiff output on my example edit of the swarm sample.

R=karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org//12084083

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@17901 260f80e4-7a28-3924-810f-c04153c831b5