Fix subscription handling in stream decoder
The _StringDecoder did not handle unsubscribe correctly (the added
test did not terminate as the unsubscribe did not cause stdin to be
closed)

Fixed by extending StreamEventTransformer which has all the
logic. Changed a number of other transformers to use
StreamEventTransformer as well.

Fixed a bug in the StreamEventTransformer implementation.

R=floitsch@google.com

BUG=

Review URL: https://codereview.chromium.org//12313127

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@19046 260f80e4-7a28-3924-810f-c04153c831b5