Make sure that no decrypted text gets stuck in the plaintext buffer
When read events are enabled on a raw secure sockets, we need to fire
off a read event if there is data in the plaintext buffer. We cannot
rely on the underlying raw socket firing off another read event even
if it is open.

R=whesse@google.com
BUG=https://code.google.com/p/dart/issues/detail?id=8824

Review URL: https://codereview.chromium.org//12385024

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@19247 260f80e4-7a28-3924-810f-c04153c831b5