Properly exposing NodeIterator and TreeWalker
Also removing references to EntityReference which is an outdated XML type.

Note that there's a bit of an oddity here around NodeFilter, which is both a class with all of the filter flags, and a function signature for a callback.

For now, I've scrubbed references to NodeFilter (the callback), but if desired I can add a new filter callback. This will most likely require a bit of Dartium work as well, so a separate CL I think.

BUG=
R=jmesserly@google.com

Review URL: https://codereview.chromium.org//13685021

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@22160 260f80e4-7a28-3924-810f-c04153c831b5