Dartc mocking and decoupling.
Herein lies a mocking of dartc.  It's close to minimal and all but test plugins have been updated to depend on it rather than the real deal.

Moving ahead tests need to be cleaned up so that we can remove dartc from the plugin build altogether.

R=brianwilkerson@google.com, danrubel@google.com, scheglov@google.com

Review URL: https://codereview.chromium.org//15679011

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@23359 260f80e4-7a28-3924-810f-c04153c831b5